refactor: remove unused presenter files and reorganize repository interfaces
- Deleted CustomerPresenter, OrderPresenter, PaymentPresenter, and ProductPresenter files as they were no longer needed.
- Moved repository interfaces from domain to application layer for better organization.
- Updated test files to reflect new repository interface paths.
- Added new repository interface files for Category, Customer, Order, and Product with CRUD methods.
- Implemented UUIDService interface for generating and validating UUIDs.
- Refactored value objects for CPF and Phone with validation and formatting logic.
- Created API routes for Category, Customer, Order, and Product with OpenAPI documentation.
- Implemented controllers for handling requests and responses, utilizing presenters for formatting output.
