+++ src/application/repositories/IOrderRepository.ts
@@ -1,4 +1,4 @@
-import Order from "../entities/Order";
+import Order from "../../domain/entities/Order";
 
 export interface IOrderRepository {
   create(order: Order): Promise<Order>;
