+++ src/domain/services/UUIDService.ts
@@ -1,0 +1,4 @@
+export interface UUIDService {
+  generate(): string;
+  validate(id: string): boolean;
+}
