+++ src/infrastructure/api/CategoryApi.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { CategoryController } from "@controllers/CategoryController";
+import { CategoryController } from "../../presentation/controllers/CategoryController";
 
 /**
  * @openapi
