+++ src/infrastructure/api/ProductApi.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { ProductController } from "@controllers/ProductController";
+import { ProductController } from "../../presentation/controllers/ProductController";
 
 /**
  * @openapi
